Separate Api in a client-api and network Requester
