+++ src/method.ts
@@ -1,0 +1,15 @@
+export { type Method };
+
+/**
+ * type for http request methods
+ */
+type Method =
+    | "CONNECT"
+    | "DELETE"
+    | "GET"
+    | "HEAD"
+    | "OPTIONS"
+    | "PATCH"
+    | "POST"
+    | "PUT"
+    | "TRACE";
